Created Proxy response class and using it instead of copypaste
